On reply to the issue #85
Added ability to change the text of the "Blank" widget to a custom one.
The rendering has not been deliberately changed to maintain compatibility with the existing layout when using with default config.
I propose to add auto sizing here as well

I used the displayName configuration parameter to not add new parameters and keep file modifications to a minimum