Match directionTrue instead of direction
It seems that when in SignalK simulator mode the send out value is "directionTrue" instead of "direction" (also in the specs).

This resulted in the true wind angle not changing in at least simulation mode. Probably on boats the other properties are used (TrueWindAngle) so this issue does not appear there.